% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2bib.r
\docType{methods}
\name{doi2bib}
\alias{doi2bib}
\alias{doi2bib,character,character-method}
\alias{doi2bib,character,missing-method}
\title{Convert doi to bibtex}
\usage{
doi2bib(doi, citekey)

\S4method{doi2bib}{character,missing}(doi, citekey)

\S4method{doi2bib}{character,character}(doi, citekey)
}
\arguments{
\item{doi}{The character string of the doi.}

\item{citekey}{Character string of equal length to \code{doi}. If used will
replace citekeys with these values.}
}
\value{
a bibtex entry as a character string.
}
\description{
Convert a digital object identifier (doi) string into a bibtex entry using
the webservice \url{http://www.doi2bib.org}.
}
\section{Methods (by class)}{
\itemize{
\item \code{doi = character,citekey = missing}: Convert doi to bibtex

\item \code{doi = character,citekey = character}: Convert doi to bibtex
}}
\examples{
doi2bib("10.1038/35012251")

}

