% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2bib.r
\docType{methods}
\name{doi2bib}
\alias{doi2bib}
\alias{doi2bib,character-method}
\title{Convert doi to bibtex}
\usage{
doi2bib(..., file, quiet = FALSE)

\S4method{doi2bib}{character}(..., file, quiet = FALSE)
}
\arguments{
\item{...}{One or more character strings of the doi. If arguments are named, names will replace citekeys}

\item{file}{an optional character string. If used the bibtex references are sent to file rather than returned.}

\item{quiet}{logical. Should the bibtex references be printed to the console?}
}
\value{
a list, returned invisibly, of bibtex references as a character strings, unless a file is specified, in which case nothing is returned
}
\description{
Convert a digital object identifier (doi) string into a bibtex entry using
the webservice \url{http://www.doi2bib.org}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Convert doi to bibtex
}}
\examples{
doi2bib(Margules2000 = "10.1038/35012251")

}

