% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2bib.r
\name{doi2bib}
\alias{doi2bib}
\title{Convert doi to bibtex}
\usage{
doi2bib(doi, citekey)
}
\arguments{
\item{doi}{The character string of the doi.}

\item{citekey}{Character string of equal length to \code{doi}. If used will replace citekeys with these values.}
}
\value{
a bibtex entry as a character string.
}
\description{
Convert a digital object identifier (doi) string into a bibtex entry using the
webservice \link{http://www.doi2bib.org}.
}
\examples{
doi2bib("10.1038/35012251")

}

